+++ resources/js/components/ProgressModal.vue
@@ -1,6 +1,7 @@
 <template>
     <sui-modal size="tiny" v-model="visible">
         <sui-modal-header>{{ title }}</sui-modal-header>
+        <sui-progress attached top :percent="done" />
         <sui-modal-content>
             <sui-list>
                 <sui-list-item v-for="step in steps" :key="step.name">
@@ -20,6 +21,7 @@ import { mapGetters } from 'vuex';
 export default {
     computed: {
         ...mapGetters({
+            done: 'progress/done',
             title: 'progress/title',
             steps: 'progress/steps',
             visible: 'progress/visible',