+++ resources/js/store/modules/Progress.js
@@ -2,6 +2,7 @@ export default {
     namespaced: true,
     state: {
         isVisible: false,
+        percentComplete: 5,
         steps: [],
         title: 'Loading...',
     },
@@ -13,6 +14,9 @@ export default {
             const index = state.steps.findIndex(s => s.name === step);
 
             Vue.set(state.steps, index, { ...state.steps[index], icon: 'check' });
+        },
+        setProgress: (state, progress) => {
+            state.percentComplete = progress;
         },
         setTitle: (state, title) => {
             state.title = title;
@@ -29,11 +33,13 @@ export default {
 
             commit('setVisible', true);
         },
-        progress: ({ commit }, step) => {
+        progress: ({ commit }, { step, progress }) => {
             commit('completeStep', step);
+            commit('setProgress', progress);
         },
     },
     getters: {
+        done: state => state.percentComplete,
         title: state => state.title,
         steps: state => state.steps,
         visible: state => state.isVisible,