+++ resources/js/store/modules/Project.js
@@ -52,20 +52,29 @@ export default {
                 const newProject = response.data,
                     projectUri = `/api/projects/${newProject.id}`;
 
-                dispatch('progress/progress', 'create', { root: true });
+                dispatch('progress/progress', {
+                    step: 'create',
+                    progress: 40,
+                }, { root: true });
 
                 if ('applications' in data) {
                     axios.post(`${projectUri}/apps`, data.applications[0]).then(appRes => {
                         newProject.applications = [ appRes ];
                         commit('addNewProject', newProject);
-                        dispatch('progress/progress', 'app', { root: true });
+                        dispatch('progress/progress', {
+                            step: 'app',
+                            progress: 100,
+                        }, { root: true });
                         resolve(response);
                     }).catch(error => reject(error));
                 } else if ('redirects' in data) {
                     axios.post(`${projectUri}/redirects`, data.redirects[0]).then(redirectRes => {
                         newProject.redirects = [ redirectRes ];
                         commit('addNewProject', newProject);
-                        dispatch('progress/progress', 'redirect', { root: true });
+                        dispatch('progress/progress', {
+                            step: 'redirect',
+                            progress: 100,
+                        }, { root: true });
                         resolve(response);
                     }).catch(error => reject(error));
                 }